Move file AddingContent + edit .mocharc.js
