+++ .mocharc.js
@@ -7,16 +7,16 @@ module.exports = {
     file: [
         './TestCases/TestBase.js',
 
-        './Testcases/monitor/backend/Login.js',
-        './Testcases/monitor/backend/Papp.js',
-        './Testcases/monitor/coinService/Balance.js',
-        './Testcases/monitor/coinService/Liquidity.js',
-        './Testcases/monitor/coinService/Market.js',
-        './Testcases/monitor/coinService/Pdex.js',
-        './Testcases/webBaseV1/BE_EstimateSwap.js',
-        './Testcases/webBaseV1/BE_EvmShield.js',
-        './Testcases/webBaseV1/BE_EvmUnshield.js',
-        './Testcases/webBaseV1/BE_SwapTxHistory.js',
+        './TestCases/monitor/backend/Login.js',
+        './TestCases/monitor/backend/Papp.js',
+        './TestCases/monitor/coinService/Balance.js',
+        './TestCases/monitor/coinService/Liquidity.js',
+        './TestCases/monitor/coinService/Market.js',
+        './TestCases/monitor/coinService/Pdex.js',
+        './TestCases/webBaseV1/BE_EstimateSwap.js',
+        './TestCases/webBaseV1/BE_EvmShield.js',
+        './TestCases/webBaseV1/BE_EvmUnshield.js',
+        './TestCases/webBaseV1/BE_SwapTxHistory.js',
     ],
     reporter: 'node_modules/mochawesome',
     'reporter-option': [
