+++ TestCases/TestBase.js
@@ -1,5 +1,5 @@
 const addContext = require('mochawesome/addContext');
-const addingContent = require("./../testbase/addingContent");
+const addingContent = require("../lib/Utils/AddingContent");
 
 beforeEach(function() {
     addingContent.resetContent()
