+++ lib/Incognito/WebServiceApi.js
@@ -1,5 +1,5 @@
 const { BaseApi } = require("../Base/BaseApi");
-const addingContent = require("../../testbase/addingContent");
+const addingContent = require("../Utils/AddingContent");
 
 class WebServiceApi extends BaseApi {
