+++ lib/Utils/AddingContent.js
